requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
bcrypt==4.2.1
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-socketio==0.0.10
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pyjwt==2.10.1
pyparsing==3.2.0
pytest==8.3.5
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.2
tomli==2.2.1
tqdm==4.66.6
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.32.0
websockets==14.2
wsproto==1.2.0
